chore: avoid using typeless binders in structure fields that shadow `variable` parameters (#23469)
This is motivated by https://github.com/leanprover/lean4/pull/7742, which makes it possible to override the inferred binder kinds of structure/class fields using typeless binders. These are all cases where field binders shadow a possible parameter for the type. The fix is to give the binder a type.