refactor(LinearAlgebra/AffineSpace/AffineSubspace): `affineSpan` analogues of `spanPoints` lemmas (#23255)
The `spanPoints` definition is only meant to be used for setting up `affineSpan`; once the bundled `affineSpan` is set up, that's what should be used.

Ensure that all `spanPoints` lemmas have `affineSpan` analogues, and move users to using `affineSpan` as far as possible (in particular, all uses outside
`Mathlib/LinearAlgebra/AffineSpace/AffineSubspace/Defs.lean` now use `affineSpan`, where previously some used `spanPoints` lemmas relying on defeq to `affineSpan`).

This would facilitate a potential future cleanup (that I'm not planning to work on), possibly as part of general work on consistency of closure operators in mathlib, to eliminate `spanPoints` (and thus `coe_affineSpan`).  I've added a lemma `mem_affineSpan_iff_exists` that provides the mathematical content of the definition of `spanPoints` as a lemma about `affineSpan` that could remain (with a proof that would no longer be `Iff.rfl`) if `spanPoints` is removed.