feat(Geometry/Euclidean/Simplex): simplex angle properties (#21960)
Show that the angles of an equilateral simplex in a Euclidean space are `π / 3`, and define the property of a simplex (typically a triangle) being acute-angled.  (This property is more common and more useful than the alternatives of "has a right angle" and "has an obtuse angle", since when it's given that a triangle has one of the latter properties, normally it's stated *which* angle is a right angle or an obtuse angle, in which case you just have an equation or inequality for a given angle.  So while those other properties will be wanted eventually - occasionally a problem does involve such properties without stating which angle is involved - they are lower priority.)