feat(Matroid/Rank): ENat-valued matroid rank (#23377)
We define the `ENat`-valued rank `M.eRank` of a matroid `M`, and also the function `M.eRk` which gives the rank of each set in `M`. We give basic API for these functions, including submodularity. 

These are analogous to the existing cardinal-valued `Matroid.cRank` and `Matroid.cRk`, but are much nicer, because the bases of a matroid are always equicardinal in the sense of `ENat`-valued cardinality but not `Cardinal`-valued.