perf: use `singlePass := true` in main loop of `ring_nf` (#23069)
DONE: the same change for `abel_nf`.

In the current situation, ring normalization is tried again on each already normalized expression, because `simp` wants to be sure it doesn't simplify any further. `singlePass := true` bypasses this.

Besides performance, there is another reason to use `singlePass := true`:
when there are bound variables, and `simp` does multiple passes, then in the second pass, the bound variable is introduced as a different fresh variable, which confuses the `AtomM` equality of atoms used by `ring`. So I need this for #22956

I also noticed that `Context.simp` only needs a `MetaM` monad and not a `SimpM` monad.