feat(CategoryTheory/Sums/Basic): functors out of `Sum` (#22278)
We prove that the category structure defined in `CategoryTheory/Sums/Basic` satisfies a universal property that exhibits it as the binary coproduct of categories. In practive, this means that we construct an equivalence of categories `(A ⊕ A') ⥤ B ≌ (A ⥤ B) × (A' ⥤ B)`, which we characterize via its compositions in both directions with `Sum.inl_`, `Sum.inr_`, `Prod.fst` and `Prod.snd`.