feat: behavior of divisors under addition (#23518)
Deliver on two elementary TODOs: Show that the divisor of an analytic function is non-negative. Show that adding an analytic function to a meromorphic one does not change the pole divisor.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.