chore(Algebra): split `Order/Floor.lean` into `Defs` and `Basic`
This is an alternative approach to #23652: `Data.NNReal.Basic` imports `Algebra.Order.Floor` but it barely seems to be used downstream. So by splitting it into just the definitions that `Data.NNReal.Basic` needs, we would be able to save a lot of imports. Unfortunetely, we still need `Floor.Basic` in `Data.Rat.Floor` which is imported into `Algebra.Order.Archimedean.Basic` (a good splitting target!), so this PR by itself does not have impressive downstream effects.