feat(Tactic/Linter): one option for all Mathlib linters
This is intended to make downstream projects easier to maintain and keep up to Mathlib standards, by having one option `linter.mathlibStandard` control all the linters that we previously had to enable one by one in the lakefile.

This is accomplished by introducing a new check `Mathlib.getLinterValue` that takes the `linter.mathlibStandard` option into account. If that option is set to true, it checks if this is one of the Mathlib standard linters, and enables the linter if so.

I also considered a few ways of implementing this feature by setting options automatically. But I couldn't gert any to work without modifying Lean itself:
* an `initialize` command doesn't get access to the options in the current file (since it runs in IO rather than CoreM)
* hooking the `runLinters` step of elaboration doesn't work because each linter restores the state (including options) after running

A drawback of the current approach is that a linter defined upstream of Mathlib cannot be enabled this way. There is no such linter currently, and we can fall back to the previous `lakefile` approach if that happens.