feat(GroupTheory/FreeGroup/Basic): add lemmas for invRev (#23366)
Upstreamed from the [EquationalTheories](https://github.com/teorth/equational_theories) project.

This PR adds the lemmas `invRev_append` and `invRev_cons` describing the interaction of `invRev` with the typical list operations. Note that `invRev_append` is a simp lemma, but `invRev_cons` is not, as otherwise there would be an infinite simp loop.