feat(PrimeSpectrum): Clopens in spectra of CommSemirings (#20533)
We show that clopens in a commutative semiring are in 1-1 correspondence with pairs of elements `(e,f)` such that `e * f = 0` and `e + f = 1`. This happens precisely when `![e,f]` is a complete orthogonal family of idempotents. Not all idempotents in a comm. semiring has such a complement: for example, in the semiring {0, 0.5, 1} with sup as + and inf as *, 0.5 doesn't have a complement.