ci: install Elan and configure Lean via `lean-action` (#23868)
The [`lean-action`](https://github.com/leanprover/lean-action) is supposed to handle common Lean CI workflows. If we disable every other option, that action installs Elan and checks the Lean setup makes sense. We do exactly that step via a custom command in a lot of workflows, so let's replace all the copy-pasted Elan setup with a call to this action.

Of course, `lean-action` can do a lot more like getting the cache, building the project, running linters and tests, etc. In follow-up work I want to outsource more build steps to `lean-action` and/or `lean-update`.