feat(Algebra/Algebra/Spectrum): The spectrum of a product is the union of the spectra (#23734)
This PR shows that `spectrum R a = ⋃ i, spectrum R (a i)` and `spectrum R ⟨a, b⟩ = spectrum R a ∪ spectrum R b` (and likewise for the quasispectrum). These results have been placed in the new file `Algebra/Algebra/Spectrum/Pi`. This is a new folder to which I have moved basic material about the spectrum and the quasispectrum.



Co-authored-by: Frédéric Dupuis <31101893+dupuisf@users.noreply.github.com>