feat(SimpleGraph/Walk): edge sets of walks (#23946)
We introduce a definition `SimpleGraph.Walk.edgeSet (p : SimpleGraph.Walk G u v) : Set (Sym2 V)` for the edge set of a walk, and give a few very basic API lemmas. This allows us to avoid both decidability and the ugly `p.edges.toFinset.toSet` when talking about the set of edges of a walk.