fix(Algebra): `MulMemClass.mul_mem` should not be `aesop safe`
The `@[aesop safe apply]` attribute only applies to implications that preserve provability. So for `mul_mem (ha : a ∈ s) (hb : b ∈ s) : a * b ∈ s`, there are ways to prove `a * b ∈ s` while `b ∉ s`, such as when `s` is an ideal and `a ∈ s`.

So, instead give `MulMemClass.mul_mem` the unsafe attribute, and add `Ideal.mul_mem_left` and `Ideal.mul_mem_right` to the aesop with the same `unsafe` attribute. Now we can prove membership of ideals a bit more easily.