refactor: define the dual of a cone as a pointed cone We currently have two definitions of the dual of a cone: * `Set.innerDualCone : Set H → ConvexCone ℝ H`, which is a `ConvexCone` * `PointedCone.dual : PointedCone ℝ H → PointedCone ℝ H`
But in fact `Set.innerDualCone` is always pointed, and it's useful to talk about the dual of an arbitrary set as a `PointedCone`.

This PR unifies those constructions under a single function `PointedCone.innerDual : Set H → ConvexCone ℝ H`.

The `inner` in the name corresponds to the fact that we are turning a cone in an inner product space `H` into another cone in `H`, rather than a cone in `M` into a cone in its dual `N`. We will soon need the latter generality in the Toric project.

From Toric