chore: import the flexible linter in Mathlib.Init (#21794)
This enables projects downstream of mathlib to enable it project-wide: previously, a file not importing this linter (as it e.g. imported only a single mathlib file not importing this linter) would fail to build because the linter option was not recognised.

Document why the remaining linters in `Mathlib.Tactic.Linter` are not enabled in `Mathlib.Init`.
Either, they are informational (like the `minImports` linter), or not ready to be used project-wide.
The `ppRoundtrip` linter arguably covers a middle ground; while it currently has too many false positives to be usable, one could conceivably also add it to Mathlib.Init. This can happen in a future PR, if there are reasons for doing so.

See [zulip discussion](https://leanprover.zulipchat.com/#narrow/channel/270676-lean4/topic/lake.20doesn't.20find.20option/near/499315240).




Co-authored-by: grunweg <rothgami@math.hu-berlin.de>