chore: fix indentation of multi-line items in doc-strings (#24166)
Markdown syntax requires that subsequent lines of multi-line items in a list should be indented: for example, one should write
```
* show that `K.πTruncGE e : K ⟶ K.truncGE e` induces an isomorphism
  in homology in degrees in the image of `e.f`.
```
instead of
```
* show that `K.πTruncGE e : K ⟶ K.truncGE e` induces an isomorphism
in homology in degrees in the image of `e.f`.
```

If the bullet point contains several paragraphs, this leads to actual mis-renderings, such as [this one](https://leanprover-community.github.io/mathlib4_docs/Mathlib/CategoryTheory/Limits/Preserves/FunctorCategory.html): "The idea of the proof" should be part of the bullet point, but isn't. [Here](https://leanprover-community.github.io/mathlib4_docs/Mathlib/CategoryTheory/NatIso.html) is another example.

[Zulip discussion](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Style.20.3Abicycle.3A.20.3A.20indenting.20second.20lines.20in.20doc-strings/near/500682936)

This fixes about 100 doc-strings violating this. There are several hundred further doc-strings with the same issue.

This indicates that a linter catching this issue may be worthwhile.