feat(SimpleGraph): eccentricity, radius, and center (#23932)
This PR defines the eccentricity of a vertex, the radius, and the center of a simple graph, and provides an API for them. They are placed in the `Diam.lean` file because they are closely related and have a similar API. Also this PR redefines `SimpleGraph.ediam` in terms of eccentricity, which shortens some of its proofs.