chore(RegularExpressions): clarify that TODO has pending PRs
The TODO on equivalence between regular expressions and DFAs seems to be
particularly appealing to new contributors, having spawned multiple
competing PRs working towards that goal.

It is also one where a new contributor is quite likely to produce a
design that is not inline with mathlib's API design philosophy and will
require a lot of guidance, which causes the PRs to languish in review.

Reword the TODO to make it less likely that the situation gets worse.