feat(Shake): autodetect the module(s) to shake
There were still a few references to Mathlib in Shake, in particular it had Mathlib hardcoded as the default module to shake. Instead ask Lake what the default targets are.

This is a back(?)port of https://github.com/leanprover-community/batteries/pull/1205, which should probably be merged anyway regardless of which repository we want shake to end up.

(Also remove a reference to Mathlib from the help text. We still have Mathlib as an example usage, which I think is innocent enough and can stay.)