fix(CI): use Lean error message in problem-matcher-wrap
This PR fixes the missing build error message annotations in GitHub. Lean switched to a different syntax for error messages, and the action we used to annotate the commit/PR with those errors did not have support for that format. I made a fork of the action that does know how to parse Lean error messages, and use that to wrap the builds.

I also opened [a PR to the original repo](https://github.com/liskin/gh-problem-matcher-wrap/pull/21). If that gets merged, we can switch back to the upstream repo for this action.

We do not need to change the format for linter/shake steps: those still fit the previous GCC-compatible syntax.

Zulip discussion: https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/gh-problem-matcher-wrap

Fixes: #12946