feat(CI): use `lean-action`'s "reinstall transient toolchain" step (#24305)
We had a hand-coded "reinstall transient toolchain" step in a few workflows. I added this to `lean-action` so we can call it directly from that action. We can also skip the "print lean and lake versions" step since `lean-action` does that too.

This PR also upgrades the `lean-action` version to the latest `main` branch.