fix(CI): use Lean error message in problem-matcher-wrap (#24316)
This PR fixes the missing build error message annotations in GitHub. Lean switched to a different syntax for error messages, and the action we used to annotate the commit/PR with those errors did not have support for that format. I made a fork of the action that does know how to parse Lean error messages, and use that to wrap the builds.

For example, see the output of [this build](https://github.com/leanprover-community/mathlib4/actions/runs/14617507898/job/41009027153):

<img width="1078" alt="Screenshot 2025-04-23 at 13 26 32" src="https://github.com/user-attachments/assets/944da832-6089-43e4-a7cf-9e073f18c79e" />
<img width="1162" alt="Screenshot 2025-04-23 at 14 07 16" src="https://github.com/user-attachments/assets/26aebffc-ea32-45ed-8f07-b3f0e1cad4de" />

I also opened [a PR to the original repo](https://github.com/liskin/gh-problem-matcher-wrap/pull/21). If that gets merged, we can switch back to the upstream repo for this action.

We do not need to change the format for linter/shake steps: those still fit the previous GCC-compatible syntax.

Zulip discussion: https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/gh-problem-matcher-wrap

Fixes: #12946