chore(CStarAlgebra): remove `PositiveLinearMapClass` (#24300)
This PR removes the typeclass `PositiveLinearMapClass`, meant as the morphism class corresponding to `PositiveLinearMap`. This class is equivalent to a combination of `LinearMapClass` and `OrderHomClass`, and it is probably best to remove it as it mixes algebra and order; in fact, it caused performance problems while I was experimenting with completely positive maps.