chore: remove @[trans] annotations on orders (#24277)
`trans` can use the `Trans` and `IsTrans` typeclasses to determine the transitivity of `≤` and `<`. So all these general `@[trans]` tags on these lemmas are redundant (or even bad).

There is also a bug in `trans` that causes the tactic to not work with lemmas about `≤` and `<` that are tagged with `@[trans]`, so these tags weren't doing anything anyways.