feat: make `lequivProdOfRightSplitExact` universe polymorphic (#24419)
When transferring a result from category theory, the theorem at the boundary should be responsible for immediately generalizing the universes if possible. Downstream cleanup from this change is left to a future PR.

Arguably we should avoid going through category theory at all here, as the current construction fails to characterize the resulting isomorphism.