perf(aesop_cat): use `apply_rfl` in `rfl_cat` (#24458)
`rfl` is a macro that calls both `eq_rfl` and `apply_rfl` (for backwards compatability reasons). When `rfl` succeeds this is no problem. But if `rfl` fails, it takes twice as long to fail.

Since `rfl_cat` is intended to fail often, I let it call `apply_rfl` directly instead of `rfl`.