feat(scripts/lint-style): do not hardcode modules to be linted
This is a step in making `lake exe lint-style` available for other projects. We add a command line argument for which modules should be linted, and if no modules are specified, we lint the default module(s) as given in the Lakefile.

This currently still assumes that we directly import all dependencies in the main module (like it would be generated by `lake exe mk_all`). It should be possible, but rather annoying, to instead do a transitive import check.