feat(CStarAlgebra): define completely positive maps (#24359)
This PR defines completely positive maps. A linear map `φ : A₁ →ₗ[ℂ] A₂` (where `A₁` and `A₂` are C⋆-algebras) is called completely positive (CP) if applying `φ` to all entries of a k × k matrix is also positive for every `k ∈ ℕ`.

Here we provide the definition and a few basic properties: CP maps are positive, and non-unital star algebra homomorphisms are CP. A fair amount of supporting material was also added to `CStarMatrix`. The API is still very barebones and we leave e.g. composition for subsequent PRs.



Co-authored-by: Frédéric Dupuis <31101893+dupuisf@users.noreply.github.com>