feat: behavior of MeromorphicAt.order under powers and inverses, mark 2 (#24314)
Establish the behavior of `MeromorphicAt.order` when taking powers. The results are completely analogous to the existing theorems for `AnalyticAt.order`.

This PR supersedes #22870, which was closed because of a hard-to-fix merge error.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.