chore(Algebra.Basic): give `instFaithfulSMulIntOfCharZero` and `instFaithfulSMulNatOfCharZero` default priority (#24555)
The instances `instFaithfulSMulNatOfCharZero` and `instFaithfulSMulIntOfCharZero` both don't seem to satisfy the criteria for lower instance priority in the library. They have very specific types for `R` so their keys won't match to much else. 

A danger could possibly come from chaining `FaithfulSMul` with Lean trying to insert `Nat` or `Int` in the middle and finding these. But we don't have anything like that now.