fix bug for defeq proofs
This fixes a bug where some terms of the form `f.symm.symm` might
appears in the expressions the tactic returns. This was due to the fact
that a `mkExpectedTypeHint` was forgotten, and so that some def-eq
equalities like `f.symm.symm = f` where ignored by the tactic, despite
being in the simp set when simplifying the final expression.